refactor(deps-core,deps-dart,deps-composer): dedup operator-spacing helper, remove dead Registry::package_url
Extract normalize_operator_spacing/has_spaced_operator into a shared
deps-core helper instead of duplicating it across deps-dart and
deps-composer. Delete Registry::package_url, which had zero production
consumers now that hover uses EcosystemFormatter::package_url.

Closes #272
Closes #213
