+++ crates/deps-go/src/registry.rs
@@ -445,10 +445,6 @@ impl deps_core::Registry for GoRegistry {
         Box::pin(async move { Ok(vec![]) })
     }
 
-    fn package_url(&self, name: &deps_core::PackageName) -> String {
-        package_url(name.as_str())
-    }
-
     fn select_latest_matching(
         &self,
         versions: &[Box<dyn deps_core::Version>],
