+++ crates/deps-maven/src/registry.rs
@@ -825,10 +825,6 @@ impl deps_core::Registry for MavenCentralRegistry {
         })
     }
 
-    fn package_url(&self, name: &deps_core::PackageName) -> String {
-        package_url(name.as_str())
-    }
-
     fn select_latest_matching(
         &self,
         versions: &[Box<dyn deps_core::Version>],
@@ -1404,17 +1400,6 @@ mod tests {
     fn test_registry_creation() {
         let cache = Arc::new(HttpCache::new());
         let _registry = MavenCentralRegistry::new(cache);
-    }
-
-    #[test]
-    fn test_registry_package_url_trait() {
-        use deps_core::Registry;
-        let cache = Arc::new(HttpCache::new());
-        let registry = MavenCentralRegistry::new(cache);
-        assert_eq!(
-            registry.package_url(&deps_core::PackageName::new("com.example:lib")),
-            "https://central.sonatype.com/artifact/com.example/lib"
-        );
     }
 
     #[test]
