+++ crates/deps-pypi/src/registry.rs
@@ -316,10 +316,6 @@ impl deps_core::Registry for PypiRegistry {
         })
     }
 
-    fn package_url(&self, name: &deps_core::PackageName) -> String {
-        package_url(name.as_str())
-    }
-
     fn select_latest_matching(
         &self,
         versions: &[Box<dyn deps_core::Version>],
